backend/requirements.txt
openai>=1.13.3
python-dotenv==1.0.0
fal-client
asyncpg==0.28.0
httpx>=0.24.0
requests>=2.31.0
markdownify>=0.11.6
